Refactor repo structure and provide CLI
