+++ README.md
@@ -1,75 +1,25 @@
-## Python Code for Converting Julian Day to UTC Standard Time
+# Convert Julian Day
 
-The following is a brief description of the Python code for converting Julian day to UTC standard time in datetime64 format, using the decimal_to_sexagesimal and decimal_to_time modules:
+Utilities for converting Julian day numbers to standard UTC datetimes.
 
-### Module decimal_to_sexagesimal: Convert decimal time to actual minutes
+## Usage
 
-```
-def decimal_to_sexagesimal(decimal_time):
-    """
-    Converts decimal time to actual minutes
+Install the package in editable mode (optional)::
 
-    Parameters:
-    decimal_time (float): Decimal time in hours
+    pip install -e .
 
-    Returns:
-    int: Minutes after conversion
-    """
-    minutes = round((decimal_time % 1) * 60)
-    return minutes
+Convert a Julian day from the command line:
+
+```bash
+python -m convert_julian_day.cli 2451545
 ```
 
-This module takes in a decimal time in hours and converts it to actual minutes.
+Or use the API:
 
-### Module decimal_to_time: Convert decimal time to actual time main function
+```python
+from convert_julian_day import juliandate_to_datetime
 
-```
-def decimal_to_time(decimal_time):
-    """
-    Converts decimal time to actual time
-
-    Parameters:
-    decimal_time (float): Decimal time in hours
-
-    Returns:
-    str: Actual time in HH:MM format
-    """
-    hours = int(decimal_time)
-    minutes = decimal_to_sexagesimal(decimal_time)
-    time_str = f"{hours:02d}:{minutes:02d}"
-    return time_str
+print(juliandate_to_datetime(2451545))
 ```
 
-This module takes in a decimal time in hours and converts it to actual time in HH:MM format using the decimal_to_sexagesimal module.
-
-### Module juliandate_to_utc: Convert julian day to actual date time 64 format
-
-```
-import numpy as np
-
-def juliandate_to_utc(jd):
-    """
-    Converts julian day to UTC datetime64 format
-
-    Parameters:
-    jd (float): Julian day
-
-    Returns:
-    np.datetime64: UTC datetime64 format
-    """
-    j2000 = np.datetime64("2000-01-01T12:00:00")
-    days_since_j2000 = jd - 2451545.0
-    seconds_since_j2000 = days_since_j2000 * 86400
-    dt = j2000 + np.timedelta64(int(seconds_since_j2000), "s")
-    return dt
-```
-
-This module takes in a Julian day and converts it to UTC datetime64 format using the J2000 epoch. It returns the datetime64 object.
-
-Technical notes:
-
-- These modules are intended for use in Python 3.x and may not be compatible with earlier versions.
-- The code may require additional libraries or dependencies to run, such as NumPy.
-- The decimal_to_sexagesimal module assumes a 24-hour clock format and may require modification to support other formats.
-- The juliandate_to_utc module assumes a Gregorian calendar and may require modification to support other calendars.
-- The code should be thoroughly tested and validated before being used in any critical applications, and any errors or bugs should be addressed before deployment.
+The example above prints `2000-01-01 12:00:00`.
