+++ setup.cfg
@@ -1,0 +1,13 @@
+[metadata]
+name = convert_julian_day
+version = 0.1.0
+
+[options]
+packages = find:
+package_dir = =src
+
+[options.package_data]
+* = *.py
+
+[options.packages.find]
+where = src
